Rename to PropertiesPane
