+++ src/jvmMain/kotlin/io/github/fourlastor/editor/PropertiesPane.kt
@@ -1,109 +1,117 @@
 package io.github.fourlastor.editor
 
-import androidx.compose.foundation.clickable
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import io.github.fourlastor.entity.*
-import io.kanro.compose.jetbrains.expui.control.Icon
+import io.github.fourlastor.entity.Entities
+import io.github.fourlastor.entity.Entity
 import io.kanro.compose.jetbrains.expui.control.Label
 import io.kanro.compose.jetbrains.expui.control.TextField
+import io.kanro.compose.jetbrains.expui.style.areaBackground
+import io.kanro.compose.jetbrains.expui.theme.DarkTheme
 
-/**
- * Property inspector panel.
- * Displays a tree of entities, and when selected it shows the editable data for the selected entity in a form.
- */
 @Composable
 fun PropertiesPane(
+    propertyNamesListState: LazyListState,
     entities: Entities,
-    modifier: Modifier,
+    modifier: Modifier = Modifier,
     onEntityChange: (Entity) -> Unit,
-    onEntityAdd: (parentId: Long) -> Unit,
 ) {
-    var selectedEntity by remember { mutableStateOf<EntityNode?>(null) }
-    val rootNode by remember(entities) { derivedStateOf { entities.asNode() } }
-
-    val selected = selectedEntity
-    if (selected != null) {
-        PropertyEditor(
-            node = selected,
-            onEntityChange = onEntityChange,
-            onEntityAdd = onEntityAdd,
+    Column(
+        modifier = modifier.fillMaxHeight()
+            .fillMaxWidth()
+            .areaBackground()
+    ) {
+        Spacer(
+            modifier = Modifier.height(54.dp)
+                .fillMaxWidth()
         )
-    }
-}
-
-/** Displays properties of an entity in a form. */
-@Composable
-private fun PropertyEditor(node: EntityNode, onEntityChange: (Entity) -> Unit, onEntityAdd: (parentId: Long) -> Unit) {
-    Column {
-        val entity = node.entity
-        Label(entity.name, fontSize = 12.sp)
-        TransformEditor(
-            transform = entity.transform,
-            onXChange = { onEntityChange(entity.x(it)) },
-            onYChange = { onEntityChange(entity.y(it)) },
-            onRotationChange = { onEntityChange(entity.rotation(it)) }
-        )
-        when (node) {
-            is GroupNode -> GroupEditor(node, onEntityAdd)
-            is ImageNode -> ImageEditor(node)
+        LazyColumn(
+            modifier = Modifier.fillMaxWidth()
+                .padding(vertical = 4.dp),
+            verticalArrangement = Arrangement.spacedBy(4.dp),
+            state = propertyNamesListState,
+            userScrollEnabled = false,
+        ) {
+            items(
+                count = entities.entities.size,
+                key = { entities.entities[it].id }
+            ) { entityIndex ->
+                val entity = entities.entities[entityIndex]
+                Column(
+                    verticalArrangement = Arrangement.spacedBy(2.dp)
+                ) {
+                    EntityName(entity)
+                    Property(
+                        "X",
+                        entity.transform.x.toString(),
+                        { it.toFloatOrNull() ?: 0f },
+                        { onEntityChange(entity.x(it)) }
+                    )
+                    Property(
+                        "Y",
+                        entity.transform.y.toString(),
+                        { it.toFloatOrNull() ?: 0f },
+                        { onEntityChange(entity.y(it)) }
+                    )
+                    Property(
+                        "Rotation",
+                        entity.transform.rotation.toString(),
+                        { it.toFloatOrNull() ?: 0f },
+                        { onEntityChange(entity.rotation(it)) }
+                    )
+                }
+            }
         }
     }
 }
 
-/** Displays a form to edit an entity [Transform]. */
 @Composable
-private fun TransformEditor(
-    transform: Transform,
-    onXChange: (Float) -> Unit,
-    onYChange: (Float) -> Unit,
-    onRotationChange: (Float) -> Unit,
+private fun <T> Property(
+    label: String,
+    value: String,
+    validator: (String) -> T,
+    onValueChange: (T) -> Unit,
+    modifier: Modifier = Modifier,
 ) {
-    Column {
-        NumberField(transform.translation.x, onXChange, "X")
-        NumberField(transform.translation.y, onYChange, "Y")
-        NumberField(transform.rotation, onRotationChange, "Rotation")
-    }
-}
-
-@Composable
-private fun NumberField(value: Float, onValueChange: (Float) -> Unit, label: String) {
-    var text by remember(value) { mutableStateOf(value.toString()) }
-    Column(verticalArrangement = Arrangement.spacedBy(2.dp)) {
-        Label(text = label)
+    Row(
+        modifier = modifier.fillMaxWidth()
+            .height(40.dp)
+            .background(DarkTheme.Grey1)
+            .padding(horizontal = 4.dp),
+        horizontalArrangement = Arrangement.spacedBy(2.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Label("$label ")
+        var text by remember(value) { mutableStateOf(value) }
         TextField(
             value = text,
             onValueChange = { text = it },
             modifier = Modifier.onFocusChanged {
                 if (!it.hasFocus) {
-                    val newValue = text.toFloatOrNull() ?: 0f
-                    text = newValue.toString()
-                    onValueChange(newValue)
+                    onValueChange(validator(text))
                 }
-            },
-        )
-    }
-}
-
-/** Adds a + button when editing a group, to add a new entity to it. */
-@Composable
-private fun GroupEditor(group: GroupNode, onEntityAdd: (parentId: Long) -> Unit) = group.run {
-    Column(modifier = Modifier.padding(start = 12.dp)) {
-        Icon(
-            resource = "icons/add.svg",
-            modifier = Modifier
-                .clickable(onClick = { onEntityAdd(group.entity.id) }),
+            }
         )
     }
 }
 
 @Composable
-private fun ImageEditor(
-    @Suppress("UNUSED_PARAMETER") image: EntityNode, // one day
-) {
-    Label("TODO")
+private fun EntityName(entity: Entity) {
+    Box(
+        modifier = Modifier.fillMaxWidth()
+            .height(40.dp)
+            .padding(horizontal = 4.dp),
+        contentAlignment = Alignment.CenterStart,
+    ) {
+        Label(text = entity.name, fontWeight = FontWeight.Medium, fontSize = 14.sp)
+    }
 }
